+++ trunk/StExBar/src/RegHistory.cpp
@@ -1,6 +1,6 @@
 // StExBar - an explorer toolbar
 
-// Copyright (C) 2009 - Stefan Kueng
+// Copyright (C) 2009-2010 - Stefan Kueng
 
 // This program is free software; you can redistribute it and/or
 // modify it under the terms of the GNU General Public License
@@ -96,7 +96,7 @@ int CRegHistory::Load(LPCTSTR lpszSection, LPCTSTR lpszKeyPrefix)
 		}
 	} while (!sText.empty() && n < m_nMaxHistoryItems);
 
-	return m_arEntries.size();
+	return (int)m_arEntries.size();
 }
 
 bool CRegHistory::Save() const
@@ -125,4 +125,3 @@ bool CRegHistory::Save() const
 	}
 	return true;
 }
-
